separate withdraw dialog in another component as the deposit dialog
